Relocate dns package
